quad: move values to a separate file
